Use JUnit 4 TemporaryFolder rule for temp folder.
This removes the Java 7 dependency in the FileResourceTest by using
JUnit 4's TemporaryFolder rule. This also ensures that the folder
will always be deleted after running the test.

The license header in AbstractDirectoryInputStream has also been put
where it is supposed to be, and minor cleanups of imports and code
style has been done.