METAMODEL-1103: Anchor WildcardPattern to start
This makes sure the WildcardPattern anchors its test to the start of
the string, not only the end.

This comes with a minor performance penalty (quick test said 5-10% ),
but I think correctness trumps speed here.

I would have liked to compare speed with pre-compiled regexes
(`Pattern.quote()` the operand, replace `%` with `.*?` and anchor it
with `^` and `$`), but if the LIKE operand is inside a SelectItem,
that would mean compiling a new regex pattern for every row, which
would probably be _very_ expensive, so I opted not to try. There seem
to be some similar optimization of the IN operator, so I might be
wrong.

Fixes METAMODEL-1103