METAMODEL-1205: Made CassandraUnit based test skip on JDK 9+
It seems that CassandraUnit simply does not work on JDK 9+, see related
comments in the code. For now I've added a JUnit `Assume` call which
ensures that the test code is skipped when CassandraUnit fails to come
up. Eventually, when they fix CassandraUnit, we should be able to
upgrade the CassandraUnit dependency and it would presumably start
working again.